+++ .gitignore
@@ -1,5 +1,5 @@
 # Benchmarking
-benchmarking/**/*.pickle
+benchmarks/**/*.pickle
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
@@ -69,6 +69,7 @@ target/
 
 # PyCharm environment files
 .idea/
+*.pclprof
 
 # VSCode environment files
 .vscode/
@@ -88,3 +89,4 @@ pythonenv*/
 # JS dependencies
 mesa/visualization/templates/external/
 mesa/visualization/templates/js/external/
+
