+++ benchmarks/Schelling/schelling.py
@@ -1,4 +1,3 @@
-import random
 
 from mesa import Agent, Model
 from mesa.space import SingleGrid
@@ -10,7 +9,7 @@ class SchellingAgent(Agent):
     Schelling segregation agent
     """
 
-    def __init__(self, pos, model, agent_type):
+    def __init__(self, unique_id, model, agent_type):
         """
         Create a new Schelling agent.
         Args:
@@ -18,14 +17,12 @@ class SchellingAgent(Agent):
            x, y: Agent initial location.
            agent_type: Indicator for the agent's type (minority=1, majority=0)
         """
-        super().__init__(pos, model)
-        self.pos = pos
+        super().__init__(unique_id, model)
         self.type = agent_type
 
     def step(self):
         similar = 0
-        r = self.model.radius
-        for neighbor in self.model.grid.iter_neighbors(self.pos, moore=True, radius=r):
+        for neighbor in self.model.grid.iter_neighbors(self.pos, moore=True, radius=self.model.radius):
             if neighbor.type == self.type:
                 similar += 1
 
@@ -36,7 +33,7 @@ class SchellingAgent(Agent):
             self.model.happy += 1
 
 
-class SchellingModel(Model):
+class Schelling(Model):
     """
     Model class for the Schelling segregation model.
     """
@@ -63,9 +60,9 @@ class SchellingModel(Model):
         # the coordinates of a cell as well as
         # its contents. (coord_iter)
         for _cont, pos in self.grid.coord_iter():
-            if random.random() < self.density:  # noqa: S311
-                agent_type = 1 if random.random() < self.minority_pc else 0  # noqa: S311
-                agent = SchellingAgent(pos, self, agent_type)
+            if self.random.random() < self.density:
+                agent_type = 1 if self.random.random() < self.minority_pc else 0
+                agent = SchellingAgent(self.next_id(), self, agent_type)
                 self.grid.place_agent(agent, pos)
                 self.schedule.add(agent)
 
@@ -75,3 +72,16 @@ class SchellingModel(Model):
         """
         self.happy = 0  # Reset counter of happy agents
         self.schedule.step()
+
+
+if __name__ == "__main__":
+    import time
+
+    # model = Schelling(15, 40, 40, 3, 1, 0.625)
+    model = Schelling(15, 100, 100, 8, 2, 0.8)
+
+    start_time = time.perf_counter()
+    for _ in range(100):
+        model.step()
+
+    print(time.perf_counter() - start_time)
